refactor: DTO와 Command 패키지 분리
- DTO는 웹 계층(dto/), Command는 도메인 계층(port/model/)으로 분리
- 계층 간 책임 명확화 및 의존성 방향 개선
