+++ src/main/java/com/agenticcp/core/domain/cloud/dto/AccountValidationResponse.java
@@ -1,4 +1,4 @@
-package com.agenticcp.core.domain.cloud.port.model.account;
+package com.agenticcp.core.domain.cloud.dto;
 
 import com.agenticcp.core.common.logging.masking.Masked;
 import com.agenticcp.core.common.logging.masking.MaskingType;
@@ -20,7 +20,7 @@ import java.util.Map;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class AccountValidationResult {
+public class AccountValidationResponse {
     
     /**
      * 검증 성공 여부
