+++ src/main/java/com/agenticcp/core/domain/cloud/dto/VpcCreateRequest.java
@@ -1,4 +1,4 @@
-package com.agenticcp.core.domain.cloud.port.model.vpc;
+package com.agenticcp.core.domain.cloud.dto;
 
 import com.agenticcp.core.domain.cloud.entity.CloudProvider.ProviderType;
 import lombok.Builder;
@@ -7,20 +7,21 @@ import lombok.Value;
 import java.util.Map;
 
 /**
- * VPC 수정 요청 모델
+ * VPC 생성 요청 모델
  * 
- * VPC 수정을 위한 표준화된 요청 모델
+ * 모든 클라우드 프로바이더에서 VPC 생성을 위한 표준화된 요청 모델
  * 
  * @author AgenticCP Team
  * @version 1.0.0
  */
 @Value
 @Builder
-public class VpcUpdateRequest {
+public class VpcCreateRequest {
     ProviderType providerType;
     String accountScope;
     String region;
     String vpcName;
+    String cidrBlock;
     String description;
     Map<String, String> tags;
     String tenantKey;
